refactor: 로또 개수 계산 객체 명확한 네이밍으로 변경
