+++ src/main/java/domain/store/LottoMachine.java
