+++ src/test/java/domain/store/CashierTest.java
@@ -1,6 +1,6 @@
-package domain.lotto;
+package domain.store;
 
-import static domain.lotto.LottoPurchaseInfo.LOTTO_PRICE;
+import static domain.store.Cashier.LOTTO_PRICE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -8,7 +8,7 @@ import domain.money.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class LottoPurchaseInfoTest {
+class CashierTest {
 
     @Test
     @DisplayName("구입 금액이 정해진 금액 단위가 아니면 예외가 발생한다.")
@@ -16,7 +16,7 @@ class LottoPurchaseInfoTest {
         Money amount = Money.from("1100");
         int manualCount = 1;
 
-        assertThatThrownBy(() -> new LottoPurchaseInfo(amount, manualCount))
+        assertThatThrownBy(() -> new Cashier(amount, manualCount))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("구입 금액은 %s원 단위로 입력해야 합니다.".formatted(LOTTO_PRICE.amount()));
     }
@@ -27,7 +27,7 @@ class LottoPurchaseInfoTest {
         Money amount = Money.from("500");
         int manualCount = 0;
 
-        assertThatThrownBy(() -> new LottoPurchaseInfo(amount, manualCount))
+        assertThatThrownBy(() -> new Cashier(amount, manualCount))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("최소 %s원 이상 입력해야 합니다.".formatted(LOTTO_PRICE.amount()));
     }
@@ -38,7 +38,7 @@ class LottoPurchaseInfoTest {
         Money amount = Money.from("3000");
         int manualCount = 4;
 
-        assertThatThrownBy(() -> new LottoPurchaseInfo(amount, manualCount))
+        assertThatThrownBy(() -> new Cashier(amount, manualCount))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("수동 로또 개수가 전체 구매 가능한 개수보다 적어야 합니다.");
     }
@@ -49,9 +49,9 @@ class LottoPurchaseInfoTest {
         Money amount = Money.from("5000");
         int manualCount = 2;
 
-        LottoPurchaseInfo info = new LottoPurchaseInfo(amount, manualCount);
+        Cashier cashier = new Cashier(amount, manualCount);
 
-        assertThat(info.getTotalCount()).isEqualTo(5);
-        assertThat(info.getAutoCount()).isEqualTo(3);
+        assertThat(cashier.getTotalCount()).isEqualTo(5);
+        assertThat(cashier.getAutoCount()).isEqualTo(3);
     }
 }
