+++ src/test/java/domain/store/LottoMachineTest.java
